requirements.txt
# Local Repo
-e .

# External
bs4
dropbox
imblearn
joblib
jupyter
lxml
matplotlib
pandas
plotly
pybeerxml
python-dotenv
requests
sklearn
tables
wget
tqdm
seaborn
